Graphe.java : le poids est maintenant un double.
Auparavant il été un int cependant avec des cas de graphe possèdant un grande quantité de sommets et d'arête le int ne suffisait plus.

+++ src/graphe/Graphe.java
@@ -178,8 +178,8 @@ public class Graphe implements GrapheInt {
         return list_edges.size();
     }
     
-    public int getPoids(){
-        int p = 0;
+    public double getPoids(){
+        double p = 0;
         for (Iterator<Edge> iterator = list_edges.iterator(); iterator.hasNext();) {
             Edge edge = (Edge) iterator.next();
             p += edge.getPoids();
